requirements.txt
pandas==2.2.0
pyyaml==7.0
matplotlib==3.8.0
seaborn==0.12.2
